requirements.txt
altair
pandas
streamlit
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
sshtunnel==0.4.0